Phase 4: the CONF reader, o_str and o_dir, and the lhash insert order its dump exposed
The stratum's implementation and the evidence for it.

Conf reader. The whole of crypto/conf: both character-class tables transcribed
exactly (they differ at six bytes -- 34, 35, 39, 59, 92, 96 -- and the default
table marks '#' CONF_COMMENT while leaving ';' plain, while win32 marks ';'
FCOMMENT|PUNCT and '#' nothing at all), the scanner, all three variable
expansion forms, clear_comments, include processing and directory walking, the
CONF container and its hash, the classic CONF_set_nconf bridge over the same
repr(C) layout, and the NCONF_ accessors.

Core runtime surface the reader needed. crypto/o_str.c and crypto/o_dir.c were
in no phase's family -- the defect class D49 recorded for OPENSSL_INIT_* -- so
their thirteen exports were invisible to every obligation table; they are now
owned by Phase 3, and OPENSSL_DIR_read/end read struct dirent and struct stat
through a C shim so no field offset is assumed.

Five defects the courts found, three of them in code already recorded as
implemented:

  * OPENSSL_LH_insert appends at a bucket's TAIL, not its head. The authority's
    getrn returns &(last->next) on a miss, so a bucket's head is its oldest
    entry and doall visits in insertion order. Found by nine RT-CONF dump.text
    residuals, all permutations. Invisible to every earlier court, because
    RT-LHASH cannot reach the doall family (the authority faults at the NULL
    thunk). Prepending had survived three phases as an assumption (D52).
  * OPENSSL_LH_doall_arg_thunk's `thunk` argument is a per-node wrapper, not an
    iteration entry point; the old code called it once for the whole table,
    handing def_dump the table pointer instead of each CONF_VALUE (D60).
  * ERR_raise_data is formatted by BIO_vsnprintf (_dopr), not libc: %s of NULL
    renders as <NULL> (reachable through NCONF_get_number_e with a NULL name)
    and truncation becomes an EMPTY string rather than a truncated one (D53).
  * get_next_file must clear its own out-parameter after OPENSSL_DIR_end, or
    the parser's "still walking a directory?" test stays true and the next EOF
    calls in with a freed path. This was a segfault in the court (D54).
  * Six of 316 ERR raise-site function names were parsed from the wrong
    identifier -- observable, since ERR_get_error_all returns them. CONF_load
    was reported as HASH_OF and _dopr as dopr (D60).

Evidence. Sixteen differential courts, 3217 observations, no residual. The
implemented libcrypto surface is 455 exports; the Phase 4 families own 262, of
which 231 are implemented and 31 handed to five named later strata. Phases 0-4
all derive complete from forensics/phase-state.json, and no phase state is
typed.

Ledger integrity. A hand-off is now sticky: the eleven exports Phase 3 handed
to this stratum stay recorded as Phase 3's hand-off even though Phase 4
implemented them, because docs/PHASE-3-CORE-RUNTIME-SEAL.md records exactly
that and a ledger that reclaimed them would both contradict a sealed document
and double-count one piece of work. ownership_audit.py reconciles the two
ledgers, fails if any symbol is counted as implemented by two strata, and fails
if a deferring stratum's hand-off set differs from the receiving stratum's
(D57). Both strata's required-evidence module lists are now complete.

The lint debt this stratum carried is cleared: clippy is clean crate-wide, so
the diagnostics fell from 522 to 0 (D59).

Also records the RT-BIO-DEBUG capture fix: the one call whose destination is
NULL writes to stderr, and its message begins with the subject's address, so
the recorded stderr -- and therefore that court's evidence identity -- differed
on every run. setarch -R is refused in both containers, so the probe now
captures and scrubs that message itself; three re-runs return one identity and
FRF refuses to re-capture (D61).

+++ src/runtime/bio/bss_mem.rs
@@ -120,6 +120,10 @@ pub extern "C" fn BIO_s_secmem() -> *const BioMethod {
 /// A negative `len` means "the buffer is NUL-terminated, measure it"; the result
 /// is a read-only memory BIO whose `num` (the empty-read return value) is 0, so a
 /// read past the end reports EOF rather than a retryable failure.
+///
+/// # Safety
+/// `buf` must be NULL or readable: for `len >= 0`, for `len` bytes; for a
+/// negative `len`, as a NUL-terminated string. A NULL `buf` is rejected.
 #[no_mangle]
 pub unsafe extern "C" fn BIO_new_mem_buf(buf: *const c_void, len: c_int) -> *mut Bio {
     guard_ffi(ptr::null_mut(), || {
@@ -262,6 +266,8 @@ unsafe extern "C" fn mem_free(a: *mut Bio) -> c_int {
 /// # Safety
 /// `b` must be NULL or a live memory BIO.
 unsafe fn mem_buf_sync(b: *mut Bio) -> c_int {
+    // SAFETY: `b` is NULL or a live memory BIO per the caller's contract; `as_ref`
+    // makes a NULL argument a no-op rather than a dereference.
     if let Some(b) = unsafe { b.as_ref() } {
         if b.init != 0 && !b.ptr.is_null() {
             // SAFETY: `ptr` is a `BioBufMem` for this method.
@@ -291,8 +297,10 @@ unsafe extern "C" fn mem_read(b: *mut Bio, out: *mut c_char, outl: c_int) -> c_i
     let (bbm, flags) = unsafe { ((*b).ptr.cast::<BioBufMem>(), (*b).flags) };
     // SAFETY: `bbm` is this method's data.
     let bm = if flags & BIO_FLAGS_MEM_RDONLY != 0 {
+        // SAFETY: `bbm` is live and `buf` is its owned header.
         unsafe { (*bbm).buf }
     } else {
+        // SAFETY: `bbm` is live and `readp` is its read-cursor header.
         unsafe { (*bbm).readp }
     };
     // SAFETY: `b` is live.
@@ -441,7 +449,7 @@ unsafe extern "C" fn mem_ctrl(b: *mut Bio, cmd: c_int, num: c_long, arg: *mut c_
             };
         }
         BIO_C_FILE_TELL => {
-            ret = if (off as i64) > c_long::MAX as i64 {
+            ret = if (off as i64) > c_long::MAX {
                 -1
             } else {
                 off as c_long
@@ -544,6 +552,8 @@ unsafe extern "C" fn mem_gets(bp: *mut Bio, buf: *mut c_char, size: c_int) -> c_
     }
     // SAFETY: `bm` is live and `j` bytes are readable from `data`.
     let mut i = 0;
+    // SAFETY: `bm` is live and `j` bounds the scan, so each `data.add(i)` is within
+    // the readable region while `i < j`.
     unsafe {
         while i < j {
             if *(*bm).data.add(i as usize) == b'\n' as c_char {
